requirements.txt
pyzmq>=18.1.0
pyserial>=3.4
pytest>=5.1.3
wheel
twine
setuptools-scm
mypy